# docu/rest/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.2",
        "title": "Chimera API"
    },
    "host": "localhost:8080",
    "basePath": "/api/interface/v2",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/scenario/{scenarioID}/emailtask": {
            "get": {
                "description": "This method provides information about all email Tasks inside\n a given scenario.",
                "operationId": "getAllEmailTasks_",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "scenarioID",
                        "in": "path",
                        "required": true,
                        "x-is-map": false,
                        "type": "integer",
                        "format": "int64",
                        "description": "The ID of the scenario"
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "required": false,
                        "x-is-map": false,
                        "type": "string",
                        "description": "A Filter String, only mail tasks with a label containing\n                     this filter String will be returned."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of emailtasks for the scenarios.",
                        "schema": {
                            "$ref": "#/definitions/EmailtaskList"
                        }
                    },
                    "404": {
                        "description": "The scenario wasn't found.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/scenario/{scenarioID}/emailtask/{emailTaskID}": {
            "get": {
                "description": "This method provides information about an email Task.",
                "operationId": "getEmailTaskConfiguration_",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "scenarioID",
                        "in": "path",
                        "required": true,
                        "x-is-map": false,
                        "type": "integer",
                        "format": "int64",
                        "description": "The ID of the scenario model."
                    },
                    {
                        "name": "emailTaskID",
                        "in": "path",
                        "required": true,
                        "x-is-map": false,
                        "type": "integer",
                        "format": "int64",
                        "description": "The control node ID of the mail Task."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detailed information about the emailtask.",
                        "schema": {
                            "$ref": "#/definitions/EmailConfigJaxBean"
                        }
                    },
                    "404": {
                        "description": "There is no emailtask in the scenario with the given ids.",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "EmailtaskList": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "description": "IDs of all known emailtasks.",
                    "items": {
                        "type": "number",
                        "format": "integer"
                    }
                }
            }
        },
        "EmailConfigJaxBean": {
            "type": "object",
            "properties": {
                "receiver": {
                    "type": "string",
                    "description": "The receiver of the email."
                },
                "subject": {
                    "type": "string",
                    "description": "The subject of the email."
                },
                "content": {
                    "type": "string",
                    "description": "The content of the email."
                }
            }
        },
...
        "Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "description": "Description of the error."
                }
            }
        },
...
    }
}